refactor parametres + classe Partie
